--- a/.gitignore
+++ b/.gitignore
@@ -3,4 +3,6 @@ parse.js
 node_modules
 .idea
 *.html
+main.js
+
 find-regressions/**/*.js