--- a/src/CustomParser.elm
+++ b/src/CustomParser.elm
@@ -1,5 +1,5 @@
 module CustomParser exposing
-    ( Parser, DeadEnd, Problem(..), run
+    ( Parser, run
     , int, number, symbol, keyword, variable, end
     , succeed, problem, succeedLazy, lazy, map, map2, map3, map4, map5, map6, map7, map8, map9, map10, map11, ignore, andThen
     , oneOf, backtrackable, token
@@ -48,6 +48,7 @@ module CustomParser exposing
 
 import CustomParser.Advanced as A
 import Elm.Syntax.Range exposing (Location)
+import Parser
 import Set
 
 
@@ -80,8 +81,11 @@ Notice the last case! A `Parser` will chomp as much as possible and not worry
 about the rest. Use the [`end`](#end) parser to ensure you made it to the end
 of the string!
 
+Currently reuses the `elm/parser` `DeadEnd` type to report problems
+to avoid breaking changes
+
 -}
-run : Parser a -> String -> Result (List DeadEnd) a
+run : Parser a -> String -> Result (List Parser.DeadEnd) a
 run parser source =
     case A.run parser source of
         Ok a ->
@@ -93,7 +97,7 @@ run parser source =
 
 problemToDeadEnd : A.DeadEnd Problem -> DeadEnd
 problemToDeadEnd p =
-    DeadEnd p.row p.col p.problem
+    { row = p.row, col = p.col, problem = p.problem }
 
 
 {-| A parser can run into situations where there is no way to make progress.
@@ -106,10 +110,7 @@ character, I increment the `row` and set `col=1`.
 
 -}
 type alias DeadEnd =
-    { row : Int
-    , col : Int
-    , problem : Problem
-    }
+    Parser.DeadEnd
 
 
 {-| When you run into a `DeadEnd`, I record some information about why you
@@ -123,20 +124,8 @@ relatively nice parse errors, and I am excited to see those techniques applied
 elsewhere!
 
 -}
-type Problem
-    = Expecting String
-    | ExpectingInt
-    | ExpectingHex
-    | ExpectingOctal
-    | ExpectingBinary
-    | ExpectingFloat
-    | ExpectingNumber
-    | ExpectingVariable
-    | ExpectingSymbol String
-    | ExpectingKeyword String
-    | ExpectingEnd
-    | UnexpectedChar
-    | Problem String
+type alias Problem =
+    Parser.Problem
 
 
 {-| A parser that succeeds without chomping any characters.
@@ -347,7 +336,7 @@ an example usage.
 -}
 problem : String -> Parser a
 problem msg =
-    A.problem (Problem msg)
+    A.problem (Parser.Problem msg)
 
 
 {-| If you are parsing JSON, the values can be strings, floats, booleans,
@@ -464,7 +453,7 @@ token str =
 
 toToken : String -> A.Token Problem
 toToken str =
-    A.Token str (Expecting str)
+    A.Token str (Parser.Expecting str)
 
 
 {-| Parse a bunch of different kinds of numbers without backtracking. A parser
@@ -520,13 +509,13 @@ number :
     -> Parser a
 number i =
     A.number
-        { binary = Result.fromMaybe ExpectingBinary i.binary
-        , expecting = ExpectingNumber
-        , float = Result.fromMaybe ExpectingFloat i.float
-        , hex = Result.fromMaybe ExpectingHex i.hex
-        , int = Result.fromMaybe ExpectingInt i.int
-        , invalid = ExpectingNumber
-        , octal = Result.fromMaybe ExpectingOctal i.octal
+        { binary = Result.fromMaybe Parser.ExpectingBinary i.binary
+        , expecting = Parser.ExpectingNumber
+        , float = Result.fromMaybe Parser.ExpectingFloat i.float
+        , hex = Result.fromMaybe Parser.ExpectingHex i.hex
+        , int = Result.fromMaybe Parser.ExpectingInt i.int
+        , invalid = Parser.ExpectingNumber
+        , octal = Result.fromMaybe Parser.ExpectingOctal i.octal
         }
 
 
@@ -562,13 +551,13 @@ parser like this:
 int : Parser Int
 int =
     A.number
-        { binary = Err ExpectingInt
-        , expecting = ExpectingInt
-        , float = Err ExpectingInt
-        , hex = Err ExpectingInt
+        { binary = Err Parser.ExpectingInt
+        , expecting = Parser.ExpectingInt
+        , float = Err Parser.ExpectingInt
+        , hex = Err Parser.ExpectingInt
         , int = Ok identity
-        , invalid = ExpectingInt
-        , octal = Err ExpectingInt
+        , invalid = Parser.ExpectingInt
+        , octal = Err Parser.ExpectingInt
         }
 
 
@@ -586,7 +575,7 @@ operator it is afterwards.
 -}
 symbol : String -> Parser ()
 symbol str =
-    A.symbol (A.Token str (ExpectingSymbol str))
+    A.symbol (A.Token str (Parser.ExpectingSymbol str))
 
 
 {-| Parse keywords like `let`, `case`, and `type`.
@@ -614,7 +603,7 @@ be parsed as `let ters` and then wonder where the equals sign is! Check out the
 -}
 keyword : String -> Parser ()
 keyword kwd =
-    A.keyword (A.Token kwd (ExpectingKeyword kwd))
+    A.keyword (A.Token kwd (Parser.ExpectingKeyword kwd))
 
 
 
@@ -640,7 +629,7 @@ with `end` guarantees that you have successfully parsed the whole string.
 -}
 end : Parser ()
 end =
-    A.end ExpectingEnd
+    A.end Parser.ExpectingEnd
 
 
 
@@ -712,7 +701,7 @@ So this can chomp a character like `T` and produces a `()` value.
 -}
 chompIf : (Char -> Bool) -> Parser ()
 chompIf isGood =
-    A.chompIf isGood UnexpectedChar
+    A.chompIf isGood Parser.UnexpectedChar
 
 
 {-| Chomp zero or more characters if they pass the test. This is commonly
@@ -929,7 +918,7 @@ variable :
     -> Parser String
 variable i =
     A.variable
-        { expecting = ExpectingVariable
+        { expecting = Parser.ExpectingVariable
         , inner = i.inner
         , reserved = i.reserved
         , start = i.start