--- a/src/Elm/Parser/Declarations.elm
+++ b/src/Elm/Parser/Declarations.elm
@@ -412,7 +412,7 @@ parameterPatternsEqual =
                 , syntax = patternResult.syntax
                 }
             )
-            Patterns.pattern
+            Patterns.patternNotDirectlyComposing
             Layout.maybeLayout
         )