… benchmark/elm.json
… benchmark/src/Main.elm
… src/CustomParser/Advanced.elm
… src/CustomParser/Extra.elm
… src/Elm/Parser.elm
… src/Elm/Parser/Base.elm
… src/Elm/Parser/Comments.elm
… src/Elm/Parser/Expose.elm
… src/Elm/Parser/Expression.elm
… src/Elm/Parser/File.elm
… src/Elm/Parser/Imports.elm
… src/Elm/Parser/Layout.elm
… src/Elm/Parser/Modules.elm
… src/Elm/Parser/Node.elm
… src/Elm/Parser/Numbers.elm
… src/Elm/Parser/Patterns.elm
… src/Elm/Parser/Tokens.elm
… src/Elm/Parser/TypeAnnotation.elm
… src/Parser/Extra.elm
… src/ParserFast.elm
… src/ParserFast/Advanced.elm
… src/ParserFast/Extra.elm
… src/ParserWithComments.elm
… tests/Elm/Parser/BaseTest.elm
… tests/Elm/Parser/CommentTest.elm
… tests/Elm/Parser/DeclarationsTests.elm
… tests/Elm/Parser/ExpressionTests.elm
… tests/Elm/Parser/FileTests.elm
… tests/Elm/Parser/LayoutTests.elm
… tests/Elm/Parser/LetExpressionTests.elm