Contains tiny improvements to the old parser
followed by the more important conversion to a custom parser.
- before (taken from last PR): 16077.300000000745 ms to do 100 runs => 160.77300000000744 ms/run
- after initial optimizations: 15633.79999999702 ms to do 100 runs => 156.3379999999702 ms/run
- after conversion to custom parser: 12211.5 ms to do 100 runs => 122.115 ms/run

The new parser
- saves a lot of mapping after e.g. `symbol`
- saves many combinators/andThens, e.g. for ranges/locations, layout indentation checks and lookahead/lookbehind
- replaces keep/ignore pipelines to `mapN`

This should still preserve type compatibility of the package interface since we re-use `Parser.Problem`.
In terms of performance, this change opens the door for a lot of optimizations that were locked off before (e.g. removing the `context` field from the state or returning `{ row, column }` instead of a tuple)