@lue-bird 2024-08-05
Wow, I'm surprised this bug hasn't happened for folks.
If we have e.g.
```elm
function WithoutArgs anotherArgument =
    0
```
(same with let function arguments)
It previously parsed as `NamedPattern "WithoutArgs" [ VarPattern "anotherArgument" ]`.
The fix now also makes `as` pattern usage more robust while speeding it up because there's no need to check for composing.